+++ modules/bibcite_entity/src/Form/ReferenceForm.php
@@ -12,6 +12,7 @@ use Drupal\Core\Form\FormStateInterface;
  */
 class ReferenceForm extends ContentEntityForm {
 
+
   /**
    * {@inheritdoc}
    */